add copy command

+++ src/js/Interface/LayoutPanel/LayoutMenu/LayoutMenu.js
@@ -6,14 +6,15 @@ export default class LayoutMenu{
         this._paint= paint;
         this._target = target;
     }
-
     create(x, y){
+        //проверяем есть ли другие меню
         let allMenu = document.querySelectorAll(".selected-menu") || [];
         for (const it of allMenu) {
             it.remove();
         }
 
         let menu = document.createElement("div");
+        this._menu=menu;
         menu.classList.add("selected-menu");
         menu.style.top=y-15+ "px";
         menu.style.left=x-15 + "px";
@@ -21,8 +22,10 @@ export default class LayoutMenu{
         let menuItemCust = document.createElement("div");
         menuItemCust.classList.add("selected-menu__item");
         menuItemCust.textContent = "Настройки";
+
         menuItemCust.addEventListener("click", (e) => {
             this._setting();
+            this.removeFast();
         });
 
         let menuItemComb = document.createElement("div");
@@ -30,22 +33,33 @@ export default class LayoutMenu{
         menuItemComb.textContent = "Обьеденить";
         menuItemComb.addEventListener("click", (e) => {
             this._combine();
+            this.removeFast();
+        });
+
+        let menuItemCopy = document.createElement("div");
+        menuItemCopy.classList.add("selected-menu__item");
+        menuItemCopy.textContent = "Копировать";
+        menuItemCopy.addEventListener("click", (e) => {
+            this._copy();
+            this.removeFast();
         });
 
         let menuItemDel = document.createElement("div");
         menuItemDel.classList.add("selected-menu__item");
         menuItemDel.textContent = "Удалить";
         menuItemDel.addEventListener("click", (e) => {
-            alert("Удалить !");
-            this._delete();
+            if(confirm("Удалить выделенные слои ? ")){
+                this._delete();
+            }
+            this.removeFast();
         });
 
         menu.append(menuItemCust);
         menu.append(menuItemComb);
+        menu.append(menuItemCopy);
         menu.append(menuItemDel);
         menu.addEventListener("mouseleave", (e) => {
             this.remove(menu);
-
         });
         document.body.append(menu);
     }
@@ -55,7 +69,6 @@ export default class LayoutMenu{
         let id=setTimeout(() => {
             menu.remove();
         }, 500);
-
         menu.addEventListener("mousemove", (e) => {
             menu.classList.remove("selected-menu-hide");
             clearTimeout(id);
@@ -77,50 +90,59 @@ export default class LayoutMenu{
     }
 
     _combine(){
-
+        if(confirm("Обьеденить выбранные слои ?"))
+        {
         //ищем выделенные элементы
-        let layoutPanelBlock = document.querySelector(".layout-panel__content");
-        let selected = layoutPanelBlock.querySelectorAll(".selected");
-        if(selected.length<2) {
-            alert("Выберите хотя бы 2 слоя !");
-            return;
-        }
-        //получаем индексы выделенных элементов
-        let indexArr = Array.from(selected).map((value, index, array) => +value.getAttribute("data-index"));
-
-        let min = Math.min(...indexArr);
-        alert(min);
-
-        let layoutList=this._paint.getLayoutManager().getLayoutList();
-        let canvas = document.createElement("canvas");
-        canvas.height = this._paint.getCanvas().height;
-        canvas.width = this._paint.getCanvas().width;
-        let ctx= canvas.getContext("2d");
-
-        layoutList.forEach((value, index, array) => {
-            if(indexArr.includes(index)){
-                ctx.drawImage(value.getCanvas(), 0, 0);
+            let layoutPanelBlock = document.querySelector(".layout-panel__content");
+            let selected = layoutPanelBlock.querySelectorAll(".selected");
+            if(selected.length<2) {
+                alert("Выберите хотя бы 2 слоя !");
+                return;
             }
-        });
+            //получаем индексы выделенных элементов
+            let indexArr = Array.from(selected).map((value, index, array) => +value.getAttribute("data-index"));
 
+            let min = Math.min(...indexArr);
+            alert(min);
 
+            let layoutList=this._paint.getLayoutManager().getLayoutList();
+            let canvas = document.createElement("canvas");
+            canvas.height = this._paint.getCanvas().height;
+            canvas.width = this._paint.getCanvas().width;
+            let ctx= canvas.getContext("2d");
 
-        let layout = new Layout(canvas, ctx, true, this._paint.getLayoutManager());
+            layoutList.forEach((value, index, array) => {
+                if(indexArr.includes(index)){
+                    ctx.drawImage(value.getCanvas(), 0, 0);
+                }
+            });
 
-        this._paint.getLayoutManager().setLayout(layout, min);
-        console.log(indexArr);
-        indexArr.splice(indexArr.indexOf(min), 1);
-        console.log(indexArr);
-        this._paint.getLayoutManager().deleteLayouts(indexArr);
+            let layout = new Layout(canvas, ctx, true, this._paint.getLayoutManager());
 
-        this._paint.getLayoutManager().setCurrentLayout(min);
+            this._paint.getLayoutManager().setLayout(layout, min);
+            console.log(indexArr);
+            indexArr.splice(indexArr.indexOf(min), 1);
+            console.log(indexArr);
+            this._paint.getLayoutManager().deleteLayouts(indexArr);
+
+            this._paint.getLayoutManager().setCurrentLayout(min);
+        }
     }
 
     _setting(){
-        let settingModal = new LayoutSettingModal();
         let index = +this._target.getAttribute("data-index");
         let layout=this._paint.getLayoutManager().getLayoutList()[index];
 
-        settingModal.create(this._paint, layout);
+        let settingModal = new LayoutSettingModal(this._paint, layout);
+        settingModal.create();
     }
+    
+    _copy(){
+        let index = +this._target.getAttribute("data-index");
+        this._paint.getLayoutManager().copyLayout(index);
+    }
+    removeFast(){
+        this._menu.remove();
+    }
+    
 }
